requirements.txt
Flask
flask-login
flask-caching
flask-cors
gunicorn
joblib
numpy
python-dotenv
scikit-learn
sympy
pytest
authlib
groq
